pyproject.toml
[project]
name = "train-template"
version = "0.1.0"
description = "Template to quickly get started with training a brain foundation model"
readme = "README.md"
authors = [
    { name = "Andrii Zahorodnii" },
    { name = "Alexander Brady" }
]
requires-python = ">=3.11"
dependencies = [
    "h5py>=3.14.0",
    "hydra-core>=1.3.2",
    "lightning>=2.5.5",
    "psutil>=7.1.0",
    "python-dotenv>=1.1.1",
    "temporaldata>=0.1.3",
    "torch>=2.8.0",
    "transformers>=4.57.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.13.3",
]

[tool.ruff]
target-version = "py311"
line-length = 320

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]  # basics: style, flakes, isort, pyupgrade, bugbear
ignore = ["E501", "E402"]            # let the formatter handle wrapping

[tool.ruff.lint.isort]
known-first-party = ["mlni"]

[tool.mypy]
python_version = "3.11"
exclude = ["\\.venv", "env", "build", "dist"]
follow_imports = "skip"  # don't analyze third-party libs
ignore_missing_imports = true
